refactor(frontend): move OTHER_VALUE from types.ts to constants.ts
