+++ apps/frontend/src/pages/chat/components/MessageList/components/AskUserCard/constants.ts
@@ -1,0 +1,2 @@
+/** Sentinel value for the "Other" radio option in the questionnaire. */
+export const OTHER_VALUE = '__other__';
