+++ apps/frontend/src/pages/chat/components/MessageList/components/AskUserCard/hooks/useFormState.ts
@@ -1,7 +1,7 @@
 import {useCallback, useState} from 'react';
 
+import {OTHER_VALUE} from '../constants.js';
 import type {AnswerEntry, Question} from '../types.js';
-import {OTHER_VALUE} from '../types.js';
 
 export interface FormState {
   /** Selected option value per question index. */
