+++ apps/frontend/src/pages/chat/components/MessageList/components/AskUserCard/types.ts
@@ -9,6 +9,3 @@ export interface AnswerEntry {
   question: string;
   answer: string | null;
 }
-
-/** Sentinel value for the "Other" radio option in the questionnaire. */
-export const OTHER_VALUE = '__other__';
